% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef.R
\name{ef}
\alias{ef}
\title{Inventory structure}
\usage{
ef(nfr = "1.A.1.a", fuel, pollutant, returnfdb = FALSE)
}
\arguments{
\item{nfr}{nomenclature for reporting, one or many}

\item{fuel}{String for type of fuel.}

\item{pollutant}{String for pollutant.}

\item{returnfdb}{logical return directly db}
}
\value{
Return a data.table
}
\description{
Inventory structure
}
\examples{
{
  ef()
}
}
\keyword{emission}
\keyword{factors}
