% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_ipcc.R
\name{ef_ipcc}
\alias{ef_ipcc}
\title{Inventory structure}
\usage{
ef_ipcc(nfr = "1.A.1.a", fuel, pol, returnfdb = FALSE)
}
\arguments{
\item{nfr}{nomenclature for reporting, one or many}

\item{fuel}{String for type of fuel.}

\item{pol}{String for pollutant.}

\item{returnfdb}{logical return directly db}
}
\value{
Return a data.table
}
\description{
Inventory structure
}
\examples{
{
  ef_emep(tier = 1, fuel = "Natural gas", pol = "PM2.5")
}
}
\keyword{emission}
\keyword{factors}
