% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_emissions.R
\name{estimate_emissions}
\alias{estimate_emissions}
\title{Estimate Emissions}
\usage{
estimate_emissions(activity_gwh, ef, efficiency = 0.35)
}
\arguments{
\item{activity_gwh}{Numeric, energy generation in GWh.}

\item{ef}{data.table, emission factors from ef_emep or ef_ipcc.}

\item{efficiency}{Numeric, thermal efficiency of the plant (0 to 1), default 0.35.}
}
\value{
A data.table with activity, EF, and calculated emissions.
}
\description{
Calculate emissions by combining activity data (generation) with emission factors.
}
