% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory}
\alias{inventory}
\title{start inventory structure}
\usage{
inventory(path, ef = "emep", nfr = "all")
}
\arguments{
\item{path}{Character, default .}

\item{ef}{Character, default .}

\item{nfr}{String, nomenclature for ereporting, default "all"}
}
\description{
create directories
}
\examples{
\dontrun{
# do not run
}
}
