% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_ipcc.R
\name{ef_ipcc}
\alias{ef_ipcc}
\title{Emissions factors from  Intergovernmental Panel on Climate Change (IPCC)}
\usage{
ef_ipcc(nfr = "1.A.1", fuel, pol, returnfdb = FALSE)
}
\arguments{
\item{nfr}{String: nomenclature for reporting}

\item{fuel}{String for type of fuel.}

\item{pol}{String for pollutant.}

\item{returnfdb}{logical return directly db}
}
\value{
Return a data.table
}
\description{
Emissions factors from  Intergovernmental Panel on Climate Change (IPCC)
}
\examples{
{
ef_ipcc(fuel = "Natural Gas", pol = "CH4")
}
}
\keyword{emission}
\keyword{factors}
